fix float dtype

+++ bodo/tests/utils.py
@@ -1504,8 +1504,10 @@ def _test_equal(
                 bodo_out[bodo_out.columns[i]] = pd.array(
                     bodo_out[bodo_out.columns[i]], dtype=py_dtype
                 )
-                py_out[py_out.columns[i]] = py_out[py_out.columns[i]].map(
-                    lambda a: pd.NA if np.isnan(a) else a
+                py_out[py_out.columns[i]] = (
+                    py_out[py_out.columns[i]]
+                    .map(lambda a: pd.NA if np.isnan(a) else a)
+                    .astype(py_dtype)
                 )
 
             # Handle categorical mismatch in pd.read_parquet(fname, dtype_backend="pyarrow")
